Move ButtonsPagination.svelte component in ui-components
