+++ packages/ui-components/src/lib/index.ts
@@ -34,6 +34,7 @@ export { default as Heading } from './components/Heading.svelte';
 export { default as Text } from './components/Text.svelte';
 export { default as DropdownProperty } from './components/DropdownProperty.svelte';
 export { default as IconError } from './components/IconError.svelte';
+export { default as ButtonsPagination } from './components/ButtonsPagination.svelte';
 
 //Types
 export type { AppStoresInterface } from './types/appStores.ts';
